Port "npm-react-addons:*" to gulp
