+++ gulpfile.js
@@ -217,6 +217,137 @@ gulp.task('build:react-dom', function() {
   return out.pipe(gulp.dest('build'));
 });
 
+//------------------------------------------------------------------------------
+// npm-react-addons
+//------------------------------------------------------------------------------
+
+var addons = {
+  CSSTransitionGroup: {
+    module: 'ReactCSSTransitionGroup',
+    name: 'css-transition-group',
+  },
+  LinkedStateMixin: {
+    module: 'LinkedStateMixin',
+    name: 'linked-state-mixin',
+  },
+  Perf: {
+    module: 'ReactDefaultPerf',
+    name: 'perf',
+  },
+  PureRenderMixin: {
+    module: 'ReactComponentWithPureRenderMixin',
+    name: 'pure-render-mixin',
+  },
+  TestUtils: {
+    module: 'ReactTestUtils',
+    name: 'test-utils',
+  },
+  TransitionGroup: {
+    module: 'ReactTransitionGroup',
+    name: 'transition-group',
+  },
+  cloneWithProps: {
+    module: 'cloneWithProps',
+    name: 'clone-with-props',
+  },
+  createFragment: {
+    module: 'ReactFragment',
+    method: 'create',
+    name: 'create-fragment',
+  },
+  shallowCompare: {
+    module: 'shallowCompare',
+    name: 'shallow-compare',
+  },
+  updates: {
+    module: 'update',
+    name: 'update',
+  },
+};
+
+gulp.task('npm-react-addons:release', function() {
+  var rawPkg = fs.readFileSync('./packages/react-addons/package.json', 'utf8');
+  var license = fs.readFileSync('./LICENSE', 'utf8');
+  var patents = fs.readFileSync('./PATENTS', 'utf8');
+
+  var dest = gulp.dest('build/packages');
+
+  Object.keys(addons).map(function(k) {
+    var info = addons[k];
+
+    var pkgData = JSON.parse(rawPkg);
+    pkgData.name = 'react-addons-' + info.name;
+
+    dest.write(new gutil.File({
+      path: pkgData.name + '/package.json',
+      contents: new Buffer(JSON.stringify(pkgData, null, 2)),
+    }));
+
+    dest.write(new gutil.File({
+      path: pkgData.name + '/index.js',
+      contents: new Buffer(
+        'module.exports = require(\'react/lib/' + info.module + '\')' +
+          (info.method ? '.' + info.method : '') +
+        ';'
+      ),
+    }));
+
+    dest.write(new gutil.File({
+      path: pkgData.name + '/README',
+      contents: new Buffer(
+        '# ' + pkgData.name + '\n\n' +
+        'This package provides the React ' + k + ' add-on. See ' +
+        'http://facebook.github.io/react/docs/addons.html for more information.\n'
+      ),
+    }));
+
+    dest.write(new gutil.File({
+      path: pkgData.name + '/LICENSE',
+      contents: new Buffer(license),
+    }));
+
+    dest.write(new gutil.File({
+      path: pkgData.name + '/PATENTS',
+      contents: new Buffer(patents),
+    }));
+  });
+
+  dest.end();
+  return dest;
+});
+
+gulp.task('npm-react-addons:pack', function(done) {
+  var pending = Object.keys(addons).length;
+
+  Object.keys(addons).forEach(function(k) {
+    var pkgName = 'react-addons-' + addons[k].name;
+    var pkgDir = 'build/packages/' + pkgName;
+
+    var buildSrc = pkgName + '-' + packageJson.version + '.tgz';
+    var buildDest = pkgDir + '.tgz';
+
+    var args = ['pack', pkgDir];
+    var opts = {stdio: 'inherit'};
+
+    function next() {
+      if (-- pending === 0) {
+        done();
+      }
+    }
+
+    spawn('npm', args, opts).on('close', function(code) {
+      if (code !== 0) {
+        throw new Error('npm-react-addons:pack failed');
+      }
+      fs.rename(buildSrc, buildDest, next);
+    });
+  });
+});
+
+//------------------------------------------------------------------------------
+// version-check
+//------------------------------------------------------------------------------
+
 gulp.task('version-check', function(done) {
   var failed = false;
   var versions = {
